load image album using coil and refactoring
