+++ app/build.gradle
@@ -53,6 +53,8 @@ android {
 
 dependencies {
 
+    implementation "io.coil-kt:coil:2.3.0"
+
     def hilt_version = "2.44.2"
     implementation "com.google.dagger:hilt-android:$hilt_version"
     kapt "com.google.dagger:hilt-compiler:$hilt_version"
@@ -73,7 +75,7 @@ dependencies {
 
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.0"
 
-    def fragment_version = "1.5.6"
+    def fragment_version = "1.5.7"
     
     implementation "androidx.fragment:fragment-ktx:$fragment_version"
