+++ app/src/main/AndroidManifest.xml
@@ -16,11 +16,11 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".MainActivity">
+            android:name=".player.recentlyPlayedTracks.ui.MainActivity">
         </activity>
 
         <activity
-            android:name=".ui.auth.AuthorizationActivity"
+            android:name=".authorization.AuthorizationActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
