--- app/src/main/java/com/rodcollab/moftens/MainActivity.kt
@@ -1,21 +1,0 @@
-package com.rodcollab.moftens
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.rodcollab.moftens.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-    }
-
-}
