+++ app/src/main/java/com/rodcollab/moftens/core/model/Artist.kt
@@ -1,0 +1,6 @@
+package com.rodcollab.moftens.core.model
+
+data class Artist(
+    val id: String,
+    val name: String
+)
