+++ app/src/main/java/com/rodcollab/moftens/core/model/User.kt
@@ -1,0 +1,9 @@
+package com.rodcollab.moftens.core.model
+
+data class User(
+    val id: String,
+    val display_name: String,
+    val email: String,
+    val birthdate: String,
+    val country: String,
+)
