+++ app/src/main/java/com/rodcollab/moftens/core/prefs/Preferences.kt
@@ -1,0 +1,11 @@
+package com.rodcollab.moftens.core.prefs
+
+interface Preferences {
+
+    fun saveAuthToken(response: String)
+    fun saveUserInformation(userId: String?)
+
+    fun getAuthToken() : String?
+
+
+}
