+++ app/src/main/java/com/rodcollab/moftens/core/service/user/UserService.kt
@@ -1,0 +1,8 @@
+package com.rodcollab.moftens.core.service.user
+
+import com.rodcollab.moftens.core.model.User
+
+interface UserService {
+
+    suspend fun get(callBack: (user: User?) -> Unit)
+}
