+++ app/src/main/java/com/rodcollab/moftens/core/util/Constants.kt
@@ -1,0 +1,10 @@
+package com.rodcollab.moftens.core.util
+
+object Constants {
+
+    const val SPOTIFY_PREFS = "SPOTIFY_PREFS"
+    const val AUTH_TOKEN = "AUTH_TOKEN"
+    const val USER_ID = "USER_ID"
+    const val REQUEST_CODE = 1337
+    const val SCOPES = "user-read-recently-played,user-library-modify,user-read-email,user-read-private"
+}
