--- app/src/main/java/com/rodcollab/moftens/data/Preferences.kt
@@ -1,11 +1,0 @@
-package com.rodcollab.moftens.data
-
-interface Preferences {
-
-    fun saveAuthToken(response: String)
-    fun saveUserInformation(userId: String?)
-
-    fun getAuthToken() : String?
-
-
-}
