--- app/src/main/java/com/rodcollab/moftens/data/model/Song.kt
@@ -1,9 +1,0 @@
-package com.rodcollab.moftens.data.model
-
-import com.rodcollab.moftens.domain.Artist
-
-data class Song(
-    val id: String,
-    val name: String,
-    val artists: List<Artist>
-)
