--- app/src/main/java/com/rodcollab/moftens/data/model/User.kt
@@ -1,9 +1,0 @@
-package com.rodcollab.moftens.data.model
-
-data class User(
-    val id: String,
-    val display_name: String,
-    val email: String,
-    val birthdate: String,
-    val country: String,
-)
