--- app/src/main/java/com/rodcollab/moftens/data/service/user/UserService.kt
@@ -1,8 +1,0 @@
-package com.rodcollab.moftens.data.service.user
-
-import com.rodcollab.moftens.data.model.User
-
-interface UserService {
-
-    suspend fun get(callBack: (user: User?) -> Unit)
-}
