+++ app/src/main/java/com/rodcollab/moftens/di/AppModule.kt
@@ -2,9 +2,9 @@ package com.rodcollab.moftens.di
 
 import android.app.Application
 import android.content.SharedPreferences
-import com.rodcollab.moftens.util.Constants.SPOTIFY_PREFS
-import com.rodcollab.moftens.data.DefaultPreferences
-import com.rodcollab.moftens.data.Preferences
+import com.rodcollab.moftens.core.util.Constants.SPOTIFY_PREFS
+import com.rodcollab.moftens.core.prefs.DefaultPreferences
+import com.rodcollab.moftens.core.prefs.Preferences
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
