+++ app/src/main/java/com/rodcollab/moftens/player/recentlyPlayedTracks/domain/usecase/GetRecentlyPlayedTracksUseCase.kt
@@ -1,0 +1,8 @@
+package com.rodcollab.moftens.player.recentlyPlayedTracks.domain.usecase
+
+import com.rodcollab.moftens.player.recentlyPlayedTracks.model.SongItem
+
+interface GetRecentlyPlayedTracksUseCase {
+    suspend operator fun invoke() : List<SongItem>
+}
+
