+++ app/src/main/java/com/rodcollab/moftens/player/recentlyPlayedTracks/model/SongItem.kt
@@ -1,0 +1,8 @@
+package com.rodcollab.moftens.player.recentlyPlayedTracks.model
+
+data class SongItem(
+    val id: String,
+    val name: String,
+    val artist: String,
+    val url: String
+)
