+++ app/src/main/java/com/rodcollab/moftens/player/recentlyPlayedTracks/ui/MainActivity.kt
@@ -1,0 +1,21 @@
+package com.rodcollab.moftens.player.recentlyPlayedTracks.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.rodcollab.moftens.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+    }
+
+}
