--- app/src/main/java/com/rodcollab/moftens/util/Constants.kt
@@ -1,12 +1,0 @@
-package com.rodcollab.moftens.util
-
-import com.rodcollab.moftens.BuildConfig
-
-object Constants {
-
-    const val SPOTIFY_PREFS = "SPOTIFY_PREFS"
-    const val AUTH_TOKEN = "AUTH_TOKEN"
-    const val USER_ID = "USER_ID"
-    const val REQUEST_CODE = 1337
-    const val SCOPES = "user-read-recently-played,user-library-modify,user-read-email,user-read-private"
-}
