+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.recentlyPlayedTracks.RecentlyPlayedTracksFragment">
+    tools:context=".player.recentlyPlayedTracks.ui.view.RecentlyPlayedTracksFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
