+++ app/src/main/res/layout/item.xml
@@ -7,6 +7,7 @@
     android:paddingTop="@dimen/simple_margin"
     android:paddingBottom="@dimen/simple_margin">
 
+
     <TextView
         android:id="@+id/song_name"
         android:layout_width="0dp"
@@ -14,7 +15,7 @@
         android:layout_marginStart="@dimen/horizontal_margin"
         android:textAppearance="@style/TextAppearance.AppCompat.Medium"
         android:textStyle="bold"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintStart_toEndOf="@id/imageView"
         app:layout_constraintTop_toTopOf="parent"
         tools:text="title" />
 
@@ -24,7 +25,19 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/horizontal_margin"
         android:textStyle="italic"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintStart_toEndOf="@id/imageView"
         app:layout_constraintTop_toBottomOf="@+id/song_name"
         tools:text="subtitle" />
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
+        android:layout_marginStart="@dimen/horizontal_margin"
+        app:layout_constraintEnd_toStartOf="@id/song_name"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:srcCompat="@tools:sample/avatars"
+        android:contentDescription="@string/track_image" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
