+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
     <string name="app_name">moftens</string>
+    <string name="track_image">Track Image</string>
+    <string name="login">Login</string>
 </resources>
